# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/hotels": {
            "get": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Get list of hotels",
                "description": "Returns list of hotels",
                "operationId": "getHotelsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HotelResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Store new hotel",
                "description": "Returns hotel data",
                "operationId": "storeHotel",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HotelRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Hotel"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/hotels/{id}": {
            "get": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Get hotel information",
                "description": "Returns hotel data",
                "operationId": "getHotelById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Hotel id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Hotel"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Update existing hotel",
                "description": "Returns updated hotel data",
                "operationId": "updateHotel",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Hotel id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HotelRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Hotel"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Delete existing hotel",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteHotel",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Hotel id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/hotels/{searchKeword}/search": {
            "get": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Get searched hotel",
                "description": "Returns hotel data",
                "operationId": "getSearchedHotel",
                "parameters": [
                    {
                        "name": "searchKeword",
                        "in": "path",
                        "description": "Keyword ",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Hotel"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "HotelRequest": {
                "title": "Store Hotel request",
                "description": "Store Hotel request body data",
                "required": [
                    "name"
                ],
                "type": "object"
            },
...
            "Hotel": {
                "title": "Hotel",
                "description": "Hotel model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name of the new hotel",
                        "type": "string",
                        "example": "A nice hotel"
                    },
                    "rating": {
                        "title": "Rating",
                        "description": "Rating of the new hotel",
                        "type": "integer",
                        "example": "This is new hotel's description"
                    },
                    "category": {
                        "title": "Category",
                        "description": "Category of the new hotel",
                        "type": "string",
                        "example": "This is new hotel's description"
                    },
                    "reputation": {
                        "title": "Reputation",
                        "description": "Reputation of the new hotel",
                        "type": "integer",
                        "example": "700"
                    },
                    "price": {
                        "title": "Price",
                        "description": "Price of the new hotel",
                        "type": "integer",
                        "example": "This is new hotel's description"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "deleted_at": {
                        "title": "Deleted at",
                        "description": "Deleted at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Hotel"
                }
            },
...
            "HotelResource": {
                "title": "HotelResource",
                "description": "Hotel resource",
                "xml": {
                    "name": "HotelResource"
                }
            },
...
        }
    },
    "security": [
        []
    ]
}